% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.r
\name{adjust_known}
\alias{adjust_known}
\title{Adjust for Known Drift}
\usage{
adjust_known(x, mask, drift)
}
\arguments{
\item{x}{A vector of data with missing values.}

\item{mask}{A logical mask that identifies a subgroup of \code{x}
to correct for drift. Values outside the subgroup are ignored.
If no mask is specified,
no overall trend is assumed.}

\item{drift}{The drift value to correct for. The drift
is assumed to occur at the last element of the masked vector
\code{x} and linearly shrink to zero at the first element of
the masked vector \code{x}.}
}
\value{
The vector \code{x} with corrected values for the subgroup
specified by \code{mask}.
}
\description{
Adjust a regular time series for a known drift over a
known period. The drift is assumed to increase linearly
in time from 0 at the beginning of the time series to
the specified drift value at the end of the time series.
}
\details{
This function assumes that \code{mask} specifies a
contiguous block, and that the time series values are
equally spaced. No consideration is given to noncontiguous
masks or unequal-interval time series.
}
\examples{
x = rnorm(100, 10, 0.1) + 0.025 * (0:99)
newx = adjust_known(x, drift = 99*0.025)
if (interactive()) {
  plot(x, type = 'l')
  lines(newx, col = 'red')
}

}
