% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.r
\name{aggregate_flags}
\alias{aggregate_flags}
\title{Aggregate QAQC Flags}
\usage{
aggregate_flags(flags)
}
\arguments{
\item{flags}{a vector of QAQC Flag codes.}
}
\value{
A single representative QAQC Flag code.
}
\description{
A set of rules for handling flags during data aggregation. This
function is intended to be called within an aggregation statement,
e.g. within a call to \code{dplyr::summarize} to average 15-minute data
into hourly data, etc.
}
\details{
The following rules are applied to aggregate QAQC Flags,
in the specified order of precedence:
\itemize{
\item \code{"X"} if \emph{any} flags in the vector are \code{"X"}.
\item \code{"B"} if \emph{any} flags in the vector are \code{"B"}.
\item \code{"U"} if \emph{any} flags in the vector are \code{"U"}.
\item \code{"A"} if \emph{any} flags in the vector are \code{"A"}.
\item \code{"M"} if more than 50% of flags in the vector are \code{"M"}.
\item \code{"P"} if \emph{any} non-missing flags in the vector are \code{"P"}.
\item \code{"Q"} if \emph{all} non-missing flags in the vector are \code{"Q" or "G"}.
\item \code{"G"} if \emph{all} non-missing flags in the vector are \code{"G"}.
\item For all other cases, use the majority flag value.
}

Any result not captured by the above rules will generate an error.
}
