% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.r
\name{aggregate_tests}
\alias{aggregate_tests}
\title{Aggregate Test Results}
\usage{
aggregate_tests(..., by = c("highest", "lowest", "majority"),
  na.rm = TRUE)
}
\arguments{
\item{...}{One of more vectors of test results.}

\item{fun}{The function used to aggregate. The default \code{max} is the
most conservative approach, in that it uses the largest flag value
from the set of tests as the indicator of the overall quality of the
observation.}

\item{fun.args}{A list of additional arguments to \code{fun}. By default,
the argument \code{na.rm} is passed to \code{fun}.}
}
\value{
An integer vector of test flags.
}
\description{
Aggregate the results of multiple tests.
}
\examples{
flags = replicate(3, sample(c(1L,3L,4L), 10, replace = TRUE))
aggregate_tests(flags[,1], flags[,2], flags[,3])
aggregate_tests(flags[,1], flags[,2], flags[,3], by = "majority")

}
