% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly.r
\name{anomaly_hampel}
\alias{anomaly_hampel}
\title{Hampel Filter}
\usage{
anomaly_hampel(x, mask = !is.na(x), size = c(5L, 5L), ...)
}
\arguments{
\item{x}{A vector of data.}

\item{mask}{A logical vector that defines which values in \code{x}
will used when computing statistics. Useful when a subset
of quality-assured data is available. Default mask is non-\code{NA}
Values.}

\item{size}{A two-element integer vector specifying the number of
points before and after the current data point to include in the
moving window. Default is \code{c(5L, 5L)}, i.e., an 11-element centered
moving window.}

\item{...}{Additional arguments to pass to \code{\link[=outlier_mad]{outlier_mad()}}.}
}
\value{
A vector of flags or scores.
}
\description{
Anomaly detection using a Hampel filter, i.e., MAD outlier detection
over a moving window.
}
\examples{
x = seq(0, 34, by = 0.25)*pi
noise = rnorm(length(x), mean = 1, sd = 3)
y = sin(x) + noise
mask = noise < 1

anomaly_hampel(y, mask)

}
