% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODWGtools.r
\docType{data}
\name{bdl}
\alias{bdl}
\title{Belden's Landing Historical Data}
\format{
A data frame with 384,771 rows and 3 variables.
}
\usage{
data(bdl)
}
\description{
Historical electrical conductivity data collected at
Belden's Landing.
}
\keyword{datasets}
