% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapfill.r
\name{blocksize}
\alias{blocksize}
\title{Contiguous Blocks of Discrete Values}
\usage{
blocksize(flags)
}
\arguments{
\item{flags}{a vector of QAQC Flag codes.}
}
\value{
A vector of same length as \code{flags} where each
element is the size of the block that flag is a part of.
}
\description{
Identify contiguous periods of discrete values (i.e. QAQC Flags).
}
\examples{
flags = sample(c(1L,3L,4L), 10, replace = TRUE)
blocksize(flags)

}
