% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcstandards.r
\name{calibration_drift_error}
\alias{calibration_drift_error}
\title{Calculate Calibration Drift Error
Calculate calibration drift error from laboratory
calibration data.}
\usage{
calibration_drift_error(lab.data)
}
\arguments{
\item{lab.data}{Processed laboratory calibration data,
e.g. output from \link{read_event}.}
}
\value{
a tibble of calibration drift calculations.
}
\description{
Calculate Calibration Drift Error
Calculate calibration drift error from laboratory
calibration data.
}
