% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossing.r
\name{crossing_rate}
\alias{crossing_rate}
\title{Multivariate test}
\usage{
crossing_rate(x, y, n.dev, n.prior)
}
\arguments{
\item{x}{A vector of values.}

\item{y}{A vector of coincident observations, e.g. from a separate
sensor or nearby station.}

\item{n.dev}{The number of standard deviations to test against.}

\item{n.prior}{the number of prior observations to use for computing
the standard deviation. For example, to compute standard deviations
over the previous 25 hours from 15-minute data use \code{n.prior = 100}.}
}
\description{
Perform a multivariate rate of change test. This function can also be
used for the neighbor test. See
https://cdn.ioos.noaa.gov/media/2017/12/qartod_temperature_salinity_manual.pdf
for more information.
}
\examples{
fake.data1 = c(rnorm(10,10), rnorm(10, 50), rnorm(10, 10))
fake.data2 = rep(10,30)
crossing_rate(fake.data1, fake.data2, c(1.5, 3), 5)
crossing_rate(fake.data2, fake.data1, c(1.5, 3), 5)
crossing_rate(fake.data1, fake.data1, 1.5, 5)

}
