% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.r
\name{daily_tidal_mean}
\alias{daily_tidal_mean}
\title{Daily Tidal Mean}
\usage{
daily_tidal_mean(x, y, tide = c("high", "low"), smooth = TRUE,
  span = 0.1, family = "symmetric", degree = 2)
}
\arguments{
\item{x}{A vector of values from a single day.}

\item{y}{A vector of water surface elevations from a single day.}

\item{tide}{Return the high-tide ("high") or low-tide ("low")
average.}

\item{smooth}{If \code{TRUE}, A loess smoother will be applied to the
data prior to detecting the water surface elevation minimums
or maximums.}

\item{span}{The \code{span} argument to \code{\link[stats:loess]{stats::loess()}}. The
default span of 0.1 was found to work well with 15-minute
tidal data.}

\item{family}{The \code{family} argument to \code{\link[stats:loess]{stats::loess()}}. The
default family "symmetric" was found to work well with 15-minute
tidal data.}

\item{degree}{The \code{degree} argument to \code{\link[stats:loess]{stats::loess()}}. The
default degree of 2 was found to work well with 15-minute
tidal data.}
}
\description{
Compute the daily low-tide or high-tide average.
}
