% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.r
\name{.aggregate_factor}
\alias{.aggregate_factor}
\title{Aggregate Quality Control Factor}
\usage{
.aggregate_factor(x)
}
\arguments{
\item{x}{A vector of integers.}
}
\value{
A vector or ordered factors.
}
\description{
Convert aggregate quality control labels to ordered factor.
}
\keyword{internal}
