% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{.outlier_factor}
\alias{.outlier_factor}
\title{Outlier Factor}
\usage{
.outlier_factor(x)
}
\arguments{
\item{x}{A vector of character labels.}
}
\value{
A vector or ordered factors.
}
\description{
Convert outlier labels to ordered factor.
}
\keyword{internal}
