% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtqc.r
\name{.rtqc_factor}
\alias{.rtqc_factor}
\title{Real-time Quality Control Factor}
\usage{
.rtqc_factor(x)
}
\arguments{
\item{x}{A vector of integers.}
}
\value{
A vector or ordered factors.
}
\description{
Convert real-time quality control labels to ordered factor.
}
\keyword{internal}
