% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{ec_to_psu}
\alias{ec_to_psu}
\title{EC to PSU}
\usage{
ec_to_psu(ec)
}
\arguments{
\item{ec}{Electrical Conductivity, in microsiemens.}
}
\value{
Salinity, in PSU.
}
\description{
Convert from Electrical Conductivity (microsiemens) to to Practical
Salinity Units (PSU)
}
\details{
Conversion follows the equation described in
Wagner, R. J., Boulger Jr, R. W., Oblinger, C. J.,
& Smith, B. A. (2006). Guidelines and standard procedures for
continuous water-quality monitors: station operation, record
computation, and data reporting. No. 1-D3.
}
