% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.r
\name{flag-aggregators}
\alias{flag-aggregators}
\alias{.highest}
\alias{.lowest}
\alias{.majority}
\alias{.odwg}
\title{Flag Aggregators}
\usage{
.highest(..., na.rm)

.lowest(..., na.rm)

.majority(..., na.rm)

.odwg(..., na.rm)
}
\arguments{
\item{na.rm}{Logical indicating whether missing flags should be
ignored when calculating the aggregate flag.}

\item{m}{A matrix of flags}
}
\value{
An integer vector with same length as rows of \code{m}.
}
\description{
Helper functions for aggregating flags.
}
\section{Functions}{
\itemize{
\item \code{.highest}: Highest flag value

\item \code{.lowest}: Lowest flag value

\item \code{.majority}: Majority flag value

\item \code{.odwg}: ODWG recommended method. Uses
\code{.highest} when a flag value of 4 is present, otherwise
uses \code{.majority}.
}}

\keyword{internal}
