% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqptools.r
\docType{data}
\name{flag.descriptions}
\alias{flag.descriptions}
\title{Flag Descriptions}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 3 columns.}
\usage{
flag.descriptions
}
\description{
Table of QAQC Flags with descriptions.
}
\keyword{datasets}
\keyword{internal}
