% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.r
\name{flag_descriptions}
\alias{flag_descriptions}
\title{Flag Descriptions}
\usage{
flag_descriptions(flags)
}
\arguments{
\item{flags}{A subset of flag values to return.}
}
\description{
return a table of flag values with descriptions and notes.
}
