% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtqc.r
\name{flat_test}
\alias{flat_test}
\title{Flat line test}
\usage{
flat_test(x, rep.threshold, tol)
}
\arguments{
\item{x}{A vector of values.}

\item{rep.threshold}{A length-2 numeric vector identifying reasonable
magnitudes of difference between adjacent data points. Typically
specific to location and/or climate and based on expert judgment.}

\item{tol}{Numerical tolerance to flag observations as repeated values.}
}
\description{
Perform a flat line test. See
https://cdn.ioos.noaa.gov/media/2017/12/qartod_temperature_salinity_manual.pdf
for more information.
}
\examples{
fake.data = c(rnorm(10, 10, 2), rep(10.0, 3), rnorm(10, 10, 2), rep(7.0, 8))
flat_test(fake.data, c(4, 6), 0.01)
flat_test(fake.data, c(3, 9), 0.01)

}
