% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcstandards.r
\name{fouling_error}
\alias{fouling_error}
\title{Calculate Fouling Error}
\usage{
fouling_error(visit.data)
}
\arguments{
\item{visit.data}{Processed visit data, i.e. output
from \link{read_visit}.}
}
\value{
A tibble of fouling calculations.
}
\description{
Calculate fouling error from pre/post cleaning visit data.
}
