% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtqc.r
\name{gap_test}
\alias{gap_test}
\title{Gap Test}
\usage{
gap_test(x, inc = "15 mins", con = c("near", "less than",
  "greater than"))
}
\arguments{
\item{x}{A vector of timestamps.}

\item{inc}{a character string defining the expected time increment.}
}
\value{
An integer vector of test flags.
}
\description{
Perform a data gap test. See
https://cdn.ioos.noaa.gov/media/2017/12/qartod_temperature_salinity_manual.pdf
for more information.
}
