% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{invwhich}
\alias{invwhich}
\title{Inverse Which}
\usage{
invwhich(i, len)
}
\arguments{
\item{i}{A vector of indices.}

\item{len}{The length of the original vector.}
}
\value{
A logical vector.
}
\description{
Convert output of \code{\link[base:which]{base::which()}} to a logical vector.
}
\keyword{internal}
