% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{isofor_outliers}
\alias{isofor_outliers}
\title{Isolation Forest outlier detection}
\usage{
isofor_outliers(x, mask = !is.na(x), threshold = c(0.8, 0.9), ...)
}
\arguments{
\item{x}{A vector of data.}

\item{mask}{A logical mask that identifies a subgroup of \code{x}
to compute non-parametric statistics on. Useful when a subset
of quality-assured data is available. Default action is to
remove NA values.}

\item{threshold}{A length-two vector identifying
thresholds for "mild" and "extreme" outliers.}

\item{...}{Additional arguments to \code{isofor::iForest}, namely
\code{nt} and \code{phi}.}
}
\description{
Performs outlier detection using an Isolation Forest.
}
\details{
the values of \code{threshold} identify mild and extreme outliers
based on the Isolation Forest score in the range \code{[0,1]}.
Default values are 0.8 for "mild" outliers and 0.9 for "extreme"
outliers.
}
