% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.r
\name{list_aggregate_flags}
\alias{list_aggregate_flags}
\title{Aggregate Quality Control Flags}
\usage{
list_aggregate_flags()
}
\description{
Return a data frame of aggregate quality control flags
and ranks.
}
