% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{lof_outliers}
\alias{lof_outliers}
\title{LOF outlier detection}
\usage{
lof_outliers(
  x,
  mask = !is.na(x),
  threshold = c(1.5, 2),
  return.score = FALSE,
  ...
)
}
\arguments{
\item{x}{A vector of data.}

\item{mask}{A logical mask that identifies a subgroup of \code{x}
to compute non-parametric statistics on. Useful when a subset
of quality-assured data is available. Default action is to
ignore NA values.}

\item{threshold}{A length-two vector identifying
thresholds for "mild" and "extreme" outliers.}

\item{return.score}{if \code{TRUE}, return the numeric outlier score.
If FALSE, return a label classifying the observations as one of
"not outlier", "mild outlier", or "extreme outlier".}

\item{...}{Additional arguments to \code{dbscan::lof}, namely
\code{k}.}
}
\description{
Performs outlier detection using Local Outlier Factor algorithm.
}
\details{
the values of \code{threshold} identify mild and extreme
outliers based on the LOF score. Scores significantly larger
than 1 indicate outliers. Default values are 1.5 for "mild"
outliers and 2.0 for "extreme" outliers.
}
