% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers-multivar.r
\name{moutlier_chisq}
\alias{moutlier_chisq}
\title{Chi-squared test for multivariate outliers}
\usage{
moutlier_chisq(
  xs,
  mask = !Reduce("|", lapply(xs, is.na)),
  threshold = c(0.9, 0.95),
  return.score = FALSE
)
}
\arguments{
\item{xs}{A dataframe or list of vectors
(which will be coerced to a numeric matrix).}

\item{mask}{A logical vector that defines which values in \code{x}
will used when computing statistics. Useful when a subset
of quality-assured data is available. Default mask is non-\code{NA}
Values.}

\item{threshold}{A length-two vector identifying
thresholds for "mild" and "extreme" outliers.}

\item{return.score}{if \code{TRUE}, return the numeric outlier score.
If \code{FALSE}, return an ordered factor classifying the observations
as one of "not outlier" (1), "mild outlier" (2), or
"extreme outlier" (3).}
}
\description{
Performs a Chi-squared test for multivariate outliers.
}
\examples{
x = seq(0, 34, by = 0.25)*pi
noise = rnorm(length(x), mean = 1, sd = 3)
y = sin(x) + noise
mask = noise < 1

moutlier_chisq(list(x, y))
moutlier_chisq(list(x, y), mask)
moutlier_chisq(list(x, y), mask, threshold = c(0.8, 0.9))
moutlier_chisq(list(x, y), return.score = TRUE)

}
