% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODWGtools.r
\docType{package}
\name{ODWGtools}
\alias{ODWGtools}
\alias{ODWGtools-package}
\title{ODWG Tools}
\description{
An R package for common tasks associated with the
Outlier Detection Working Group.
}
\author{
\strong{Maintainer}: Michael Koohafkan \email{Michael.Koohafkan@water.ca.gov}

}
