% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{outlier_iforest}
\alias{outlier_iforest}
\title{Isolation Forest outlier detection}
\usage{
outlier_iforest(
  xs,
  mask = !Reduce("|", lapply(xs, is.na)),
  threshold = c(0.8, 0.9),
  return.score = FALSE,
  ...
)
}
\arguments{
\item{xs}{A dataframe or list of vectors
(which will be coerced to a numeric matrix).}

\item{mask}{A logical mask that identifies a subgroup of \code{x}
to compute statistics on. Useful when a subset
of quality-assured data is available. Default action is to
ignore NA values.}

\item{threshold}{A length-two vector identifying
thresholds for "mild" and "extreme" outliers.}

\item{return.score}{if \code{TRUE}, return the numeric outlier score.
If FALSE, return an ordered factor classifying the observations
as one of "not outlier" (1), "mild outlier" (2), or
"extreme outlier" (3).}

\item{...}{Additional arguments to \code{solitude::isolationForest$new()}.
note that the argument \code{sample_size} will be overwritten to use the
number of unmasked data points, i.e. \code{length(which(mask))}.}
}
\description{
Performs outlier detection using an Isolation Forest.
}
\details{
the values of \code{threshold} identify mild and extreme\
outliers based on the Isolation Forest score in the range \verb{[0,1]}.
Default values are 0.8 for "mild" outliers and 0.9 for "extreme"
outliers.
}
\examples{
x = seq(0, 34, by = 0.25)*pi
noise = rlnorm(length(x), meanlog = 1, sdlog = 3)
y=sin(x) + noise
mask = noise < 1

outlier_iforest(list(y))
outlier_iforest(list(x, y))
outlier_iforest(list(x, y), mask)
outlier_iforest(list(x, y), mask, threshold = c(1, 2))
outlier_iforest(list(x, y), return.score = TRUE)

}
