% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{outlier_lof}
\alias{outlier_lof}
\title{LOF outlier detection}
\usage{
outlier_lof(
  x,
  mask = !is.na(x),
  threshold = c(1.5, 2),
  return.score = FALSE,
  ...
)
}
\arguments{
\item{x}{A vector of data.}

\item{mask}{A logical mask that identifies a subgroup of \code{x}
to compute non-parametric statistics on. Useful when a subset
of quality-assured data is available. Default action is to
ignore NA values.}

\item{threshold}{A length-two vector identifying
thresholds for "mild" and "extreme" outliers.}

\item{return.score}{if \code{TRUE}, return the numeric outlier score.
If FALSE, return an ordered factor classifying the observations
as one of "not outlier" (1), "mild outlier" (2), or
"extreme outlier" (3).}

\item{...}{Additional arguments to \code{dbscan::lof}, namely
\code{k}.}
}
\description{
Performs outlier detection using Local Outlier Factor algorithm.
}
\details{
the values of \code{threshold} identify mild and extreme
outliers based on the LOF score. Scores significantly larger
than 1 indicate outliers. Default values are 1.5 for "mild"
outliers and 2.0 for "extreme" outliers.
}
\examples{
x = seq(0, 34, by = 0.25)*pi
noise = rlnorm(length(x), meanlog = 1, sdlog = 3)
y=sin(x) + noise
mask = noise < 1

outlier_lof(y)
outlier_lof(y, mask)
outlier_lof(y, mask, threshold = c(1, 2))
outlier_lof(y, return.score = TRUE)

}
