% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{outlier_tukey}
\alias{outlier_tukey}
\title{Tukey's test for outliers}
\usage{
outlier_tukey(
  x,
  mask = !is.na(x),
  threshold = c(1.5, 3),
  return.score = FALSE,
  ...
)
}
\arguments{
\item{x}{A vector of data.}

\item{mask}{A logical vector that defines which values in \code{x}
will used when computing statistics. Useful when a subset
of quality-assured data is available. Default mask is non-\code{NA}
Values.}

\item{threshold}{A length-two vector identifying
thresholds for "mild" and "extreme" outliers.}

\item{return.score}{if \code{TRUE}, return the numeric outlier score.
If \code{FALSE}, return an ordered factor classifying the observations
as one of "not outlier" (1), "mild outlier" (2), or
"extreme outlier" (3).}

\item{...}{
  Arguments passed on to \code{\link[stats:quantile]{stats::quantile}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A  vector the same length as \code{x} containing numeric
scores or ordered factors.
}
\description{
Perform Tukey's test for mild and extreme outliers.
}
\details{
the values of \code{threshold} identify the multiplier of the
interquartile range used to identify mild and extreme outliers.
typical values are 1.5 for "mild" outliers and 3.0 for "extreme"
outliers.
}
\examples{
x = seq(0, 34, by = 0.25)*pi
noise = rnorm(length(x), mean = 1, sd = 3)
y = sin(x) + noise
mask = noise < 1

outlier_tukey(y)
outlier_tukey(y, mask)
outlier_tukey(y, mask, threshold = c(2, 5))
outlier_tukey(y, return.score = TRUE)

}
