% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtqc.r
\name{range_test}
\alias{range_test}
\title{Range Test}
\usage{
range_test(x, sensor.range = c(0, 50000), user.range = c(5, 45000))
}
\arguments{
\item{x}{A vector of values.}

\item{sensor.range}{A length-2 numeric vector identifying the sensor
measurement range.}

\item{user.range}{A length-2 numeric vector identifying a reasonable
measurement range for the provided data \code{x}. Typically specific to
location and/or climate and based on expert judgment.}
}
\value{
An integer vector of test flags.
}
\description{
Perform a range test. See
https://cdn.ioos.noaa.gov/media/2017/12/qartod_temperature_salinity_manual.pdf
for more information.
}
