% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtqc.r
\name{rate_test}
\alias{rate_test}
\title{Rate test}
\usage{
rate_test(x, n.dev, n.prior)
}
\arguments{
\item{x}{A vector of values.}

\item{n.dev}{The number of standard deviations to test against.}

\item{n.prior}{the number of prior observations to use for computing
the standard deviation. For example, to compute standard deviations
over the previous 25 hours from 15-minute data use \code{ntile = 100}.}
}
\description{
Perform a rate of change test. See
https://cdn.ioos.noaa.gov/media/2017/12/qartod_temperature_salinity_manual.pdf
for more information.
}
\examples{
fake.data = c(rnorm(10,10), rnorm(10, 50), rnorm(10,10))
rate_test(fake.data, 2, 5)

}
