% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsheets.r
\name{read_calibration}
\alias{read_calibration}
\title{Read Laboratory Calibration Sheet}
\usage{
read_calibration(workbook, sheet)
}
\arguments{
\item{workbook}{The Excel workbook (.xlsx) containing
calibration sheets.}

\item{sheet}{The calibration sheet to read.}
}
\value{
A tibble of calibration data.
}
\description{
Read laboratory calibration sheet data.
}
