% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsheets.r
\name{read_event}
\alias{read_event}
\title{Read Event Sheet}
\usage{
read_event(workbook, timezone = "Etc/GMT+8")
}
\arguments{
\item{workbook}{The Excel workbook (.xlsx) containing
the event data. See 'details' for the required format.}

\item{timezone}{The timezone assumed for the visit sheet.
Default is \code{"Etc/GMT+8"} (PST).}
}
\value{
A tibble of visit data.
}
\description{
Read event data from an electronic form.
}
\details{
The workbook is expected to contain three sheets:
\itemize{
\item "RESULT": this sheet is formatted to contain
columns that match the WQP \emph{result} components:
"cdec_code", "interval_name", "analyte_name",
"unit_name", rank_name", "reading_type_name",
"time", "value".
\item "EVENT: this sheet is formatted to contain
columns that match the WQP \emph{event} components:
"cdec_code", "event_type_name", "contact_name",
"arrival_time", "restart_time", "departure_time",
"reason_name", "summary_name", "comments".
\item "ACTION": this sheet is formatted to contain columns
that match the WQP \emph{action} components:
}
}
