% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsheets.r
\name{read_visit}
\alias{read_visit}
\title{Read Visit Sheet}
\usage{
read_visit(workbook, sheet, timezone = "US/Pacific")
}
\arguments{
\item{workbook}{The Excel workbook (.xlsx) containing
visit sheets.}

\item{sheet}{The visit sheet to read.}

\item{timezone}{The timezone assumed for the visit sheet.
Default is "US/Pacific" (PST/PDT).}
}
\value{
A tibble of visit data.
}
\description{
Read field visit sheet data.
}
