% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossing.r
\name{regional_regression}
\alias{regional_regression}
\title{Regional Regression}
\usage{
regional_regression(x, y, season = NULL, plot = FALSE)
}
\arguments{
\item{x, y}{Vectors of station data used for regression.}

\item{season}{Optional vector identifying specific
seasons or periods}

\item{plot}{If \code{TRUE}, print diagnostic plots.}
}
\description{
Perform regional regression of two stations.
}
