% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtqc.r
\name{rtqc_attenuation}
\alias{rtqc_attenuation}
\title{Attenuation Test}
\usage{
rtqc_attenuation(x, threshold, n.obs, ...)
}
\arguments{
\item{x}{A vector of values.}

\item{threshold}{A length-2 vector specifying "suspect" and "fail"
thresholds for the minimum standard deviation of a set of
observations.}

\item{...}{Other arguments to pass to \code{\link[stats:sd]{stats::sd()}}, i.e.,
the argument \code{na.rm}.}
}
\value{
An ordered factor of test flags of same length as \code{x}.
}
\description{
Perform a signal attenuation test. See
https://cdn.ioos.noaa.gov/media/2017/12/qartod_temperature_salinity_manual.pdf
for more information.
}
\examples{
fake.data = map(seq(2, 0.2, by = -0.2), ~rnorm(10, sd = .x)) \%>\%
  unlist()
rtqc_attenuation(fake.data, c(0.8, 0.4), 5)

}
