% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtqc.r
\name{rtqc_flat}
\alias{rtqc_flat}
\title{Flat Line Test}
\usage{
rtqc_flat(x, rep.threshold, tol)
}
\arguments{
\item{x}{A vector of values.}

\item{rep.threshold}{A length-2 integer vector identifying reasonable
counts of repeated values before a data point is considered suspect
or indicates sensor failure. Typically
specific to location and/or climate and based on expert judgment.}

\item{tol}{Numerical tolerance to consider adjacent observations as
repeated values.}
}
\value{
An ordered factor of test flags of same length as \code{x}.
}
\description{
Perform a flat line test. For more information see
U.S. Integrated Ocean Observing System, 2015. Manual for Real-Time
Quality Control of In-situ Temperature and Salinity Data Version
2.0: A Guide to Quality Control and Quality Assurance of In-situ
Temperature and Salinity Observations. 56 pp.
DOI: \href{https://doi.org/10.7289/V5V40SD4}{10.7289/V5V40SD4}.
}
\examples{
fake.data = c(rnorm(10, 10, 2), rep(10.0, 3), rnorm(10, 10, 2),
  rep(7.0, 8))
# flag data as "suspect" when 4 observations in a row differ by
# less than 0.01, and as "fail" when 6 observations in a row
# differ by less than 0.01.
rtqc_flat(fake.data, c(4, 6), 0.01)
# as above, but 3 repeated observations signals "suspect" while
# 9 observations signals "fail".
rtqc_flat(fake.data, c(3, 9), 0.01)

}
