% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtqc.r
\name{rtqc_gap}
\alias{rtqc_gap}
\title{Gap Test}
\usage{
rtqc_gap(x, increment, condition = c("is", "less than", "greater than"))
}
\arguments{
\item{x}{A vector of timestamps.}

\item{increment}{A character string defining the expected time
increment. this consists of two parts: a numeric value defining
the size of the increment, and the units of the increment.
Acceptable units are "secs", "mins", "hours", "days", or "weeks".}

\item{condition}{A character string specifying the required
condition. "is" specifies that the interval between observations
must be exactly the supplied increment. "less than" or
"greater than" specify that the interval must be at most or at
least the specified increment, respectively.}
}
\value{
An ordered factor of test flags of same length as \code{x}.
}
\description{
Perform a data gap test. For more information see
U.S. Integrated Ocean Observing System, 2015. Manual for Real-Time
Quality Control of In-situ Temperature and Salinity Data Version
2.0: A Guide to Quality Control and Quality Assurance of In-situ
Temperature and Salinity Observations. 56 pp.
DOI: \href{https://doi.org/10.7289/V5V40SD4}{10.7289/V5V40SD4}.
}
\examples{
# Application to 15-minute data:
fake.timestamps = c(
  seq(as.POSIXct("2018-01-01"), by = "15 mins", length.out = 5),
  rep(as.POSIXct("2018-01-01 01:00:00"), 3),
  as.POSIXct("2018-01-01 04:00:00")
)
rtqc_gap(fake.timestamps, "16 mins", "less than")
rtqc_gap(fake.timestamps, "15 mins", "is")

}
