% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtqc.r
\name{rtqc_rate}
\alias{rtqc_rate}
\title{Rate Test}
\usage{
rtqc_rate(x, n.dev, n.prior, ...)
}
\arguments{
\item{x}{A vector of values.}

\item{n.dev}{The number of standard deviations to test against.}

\item{n.prior}{the number of prior observations to use for computing
the standard deviation. For example, to compute standard deviations
over the previous 25 hours from 15-minute data use \code{n.prior = 100}.}

\item{...}{Other arguments to pass to \code{\link[stats:sd]{stats::sd()}}, i.e.,
the argument \code{na.rm}.}
}
\value{
An ordered factor of test flags of same length as \code{x}.
}
\description{
Perform a rate of change test. See
https://cdn.ioos.noaa.gov/media/2017/12/qartod_temperature_salinity_manual.pdf
for more information.
}
\examples{
fake.data = c(rnorm(10,10), rnorm(10, 50), rnorm(10,10))
rtqc_rate(fake.data, 2, 5)

}
