% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtqc.r
\name{rtqc_rate_alt}
\alias{rtqc_rate_alt}
\title{Rate Test (Alternate)}
\usage{
rtqc_rate_alt(x, threshold, n.prior)
}
\arguments{
\item{x}{A vector of values.}

\item{threshold}{The absolute rate of change over a set of
observations. A rate of change above this value will flag
the final data point in the set.}

\item{n.prior}{the number of prior observations to calculate the
rate of change over, excluding the current observation.}
}
\value{
An ordered factor of test flags of same length as \code{x}.
}
\description{
Perform a rate of change test. This is a variant of \code{\link[=rtqc_flat]{rtqc_flat()}}
which checks the absolute rate of change over a given number
of observations (defined as the absolute value of the average of
the rates of change between successive observations)
}
\examples{
fake.data = c(rnorm(10,10), rnorm(10, 50), rnorm(10,10))
rtqc_rate_alt(fake.data, 20, 3)

}
