% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtqc.r
\name{rtqc_spike}
\alias{rtqc_spike}
\title{Spike test}
\usage{
rtqc_spike(x, spike.threshold)
}
\arguments{
\item{x}{A vector of values.}

\item{spike.threshold}{A length-2 numeric vector identifying
reasonable magnitudes of difference between adjacent data points.
Typically specific to location and/or climate and based on expert
judgment.}
}
\description{
Perform a spike test. See
https://cdn.ioos.noaa.gov/media/2017/12/qartod_temperature_salinity_manual.pdf
for more information.
}
\examples{
fake.data = c(rnorm(10, 10,3), 25, rnorm(10, 10,3))
rtqc_spike(fake.data, c(5, 10))

}
