% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtqc.r
\name{rtqc_spike_alt}
\alias{rtqc_spike_alt}
\title{Spike Test (Alternate)}
\usage{
rtqc_spike_alt(x, spike.threshold)
}
\arguments{
\item{x}{A vector of values.}

\item{spike.threshold}{A length-2 list specifying "suspect"
and "fail" thresholds. Each threshold can be either a single
value (static threshold) or a vector of same length as \code{x}
(dynamic threshold). Thresholds are typically specific to
location and/or climate and based on expert judgment.}
}
\value{
An ordered factor of test flags of same length as \code{x}.
}
\description{
Perform a spike test. This is a variant of \code{\link[=rtqc_spike]{rtqc_spike()}}
which checks the absolute difference between the current
measurement and the previous/next measurements.
}
\examples{
fake.data = c(1, 1.1, 1, 1.2, 1, 1.3, 1, 1.4, 1, 1.5, 1)
rtqc_spike(fake.data, c(0.15, 0.25))

}
