% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.r
\name{smooth_godin}
\alias{smooth_godin}
\title{Godin Smoother}
\usage{
smooth_godin(x, increment)
}
\arguments{
\item{x}{A vector of values.}

\item{increment}{A character string defining the expected time
increment. this consists of two parts: a numeric value defining
the size of the increment, and the units of the increment.
Acceptable units are "secs", "mins", "hours", "days", or "weeks".}
}
\value{
A vector of same length as x.
}
\description{
Apply a Godin filter to a vector of data.
}
\details{
The Godin filter is defined
\deqn{F = \frac{A_{24}^2}{24^2}\frac{A_{25}}{25}}
}
