% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{string_to_difftime}
\alias{string_to_difftime}
\title{String to difftime}
\usage{
string_to_difftime(s)
}
\arguments{
\item{s}{A string}
}
\value{
A \code{difftime} object.
}
\description{
Convert a string to a \code{difftime} object.
}
\keyword{internal}
