% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcstandards.r
\name{total_drift_error}
\alias{total_drift_error}
\title{Calculate Total Drift Error
Calculate total drift error from combined fouling and
calibration drift data.}
\usage{
total_drift_error(fouling.data, calibration.data)
}
\arguments{
\item{fouling.data}{Fouling error data,
e.g. output from \link{fouling_error}.}

\item{calibration.data}{Calibration drift error data,
e.g. output from \link{calibration_drift_error}.}
}
\value{
a tibble of total drift calculations.
}
\description{
Calculate Total Drift Error
Calculate total drift error from combined fouling and
calibration drift data.
}
