% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{tukey_outliers}
\alias{tukey_outliers}
\title{Tukey's test for outliers}
\usage{
tukey_outliers(
  x,
  mask = !is.na(x),
  threshold = c(1.5, 3),
  return.score = FALSE
)
}
\arguments{
\item{x}{A vector of data.}

\item{mask}{A logical mask that identifies a subgroup of \code{x}
to compute non-parametric statistics on. Useful when a subset
of quality-assured data is available. Default action is to
remove NA values.}

\item{threshold}{A length-two vector identifying
thresholds for "mild" and "extreme" outliers.}

\item{return.score}{if \code{TRUE}, return the numeric outlier score.
If FALSE, return a label classifying the observations as one of
"not outlier", "mild outlier", or "extreme outlier".}
}
\value{
A  vector the same length as \code{x} containing numeric
scores or character labels.
}
\description{
Perform Tukey's test for mild and extreme outliers.
}
\details{
the values of \code{threshold} identify the multiplier of the
interquartile range used to identify mild and extreme outliers.
typical values are 1.5 for "mild" outliers and 3.0 for "extreme"
outliers.
}
