% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqpformat.r
\name{visit_to_wqp}
\alias{visit_to_wqp}
\title{Visit to WQP}
\usage{
visit_to_wqp(visit.data, wqp.listings)
}
\arguments{
\item{visit.data}{Processed visit data, i.e. output
from \link{read_visit}.}

\item{wqp.listings}{A listing of result IDs from the WQP
database, e.g. output of \code{wqpr::list_all()}. Used to}
}
\value{
A tibble formatted for insertion to WQP.
}
\description{
Format visit data for insertion to WQP.
}
\seealso{
\code{\link[wqpr:list_all]{wqpr::list_all()}}
}
