% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{water_year}
\alias{water_year}
\title{Water Year}
\usage{
water_year(x)
}
\arguments{
\item{x}{A vector of DateTime values}
}
\value{
An integer vector of water years.
}
\description{
Get the water year for a particular date.
}
