% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqptools.r
\docType{package}
\name{wqptools}
\alias{wqptools}
\alias{wqptools-package}
\title{WQP Tools}
\description{
An R package for common tasks associated with the
Water Quality Portal.
}
\author{
\strong{Maintainer}: Michael Koohafkan \email{Michael.Koohafkan@water.ca.gov}

}
