% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{zscore_outliers}
\alias{zscore_outliers}
\title{Z-score test for outliers}
\usage{
zscore_outliers(x, mask = !is.na(x), threshold = c(0.9, 0.95))
}
\arguments{
\item{x}{A vector of data.}

\item{mask}{A logical mask that identifies a subgroup of \code{x}
to compute non-parametric statistics on. Useful when a subset
of quality-assured data is available. Default action is to
remove NA values.}

\item{threshold}{A length-two vector identifying
thresholds for "mild" and "extreme" outliers.}
}
\description{
Performs a Z-score test for outliers.
}
\details{
the values of \code{threshold} identify the quantiles of the
normal distribution used to identify mild and extreme outliers.
Default values are 0.9 for "mild" outliers and 0.95 for "extreme"
outliers.
}
