% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.r
\name{aggregate_tests}
\alias{aggregate_tests}
\title{Aggregate Test Results}
\usage{
aggregate_tests(
  ...,
  by = c("odwg", "highest", "lowest", "majority"),
  na.rm = TRUE
)
}
\arguments{
\item{...}{One of more vectors of test results.}

\item{by}{The method to use for aggregating. When \code{by = "highest"}
the largest flag value from the set of tests as the indicator
of the overall quality of the observation. When \code{by = "lowest"}
the smallest flag value from the set of tests as the indicator
of the overall quality of the observation. When \code{by = "majority"}
a majority-vote algorithm is used to determine the overall quality
of the observation.}

\item{na.rm}{Logical indicating whether missing flags should be
ignored when calculating the aggregate flag.}
}
\value{
An integer vector of test flags.
}
\description{
Aggregate the results of multiple tests.
}
\examples{
flags = replicate(3, sample(c(1L,3L,4L), 10, replace = TRUE))
aggregate_tests(flags[,1], flags[,2], flags[,3])
aggregate_tests(flags[,1], flags[,2], flags[,3], by = "majority")

}
