% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{chisq_outliers}
\alias{chisq_outliers}
\title{Chi-squared test for outliers}
\usage{
chisq_outliers(
  x,
  mask = !is.na(x),
  threshold = c(0.9, 0.95),
  return.score = FALSE
)
}
\arguments{
\item{x}{A vector of data.}

\item{mask}{A logical mask that identifies a subgroup of \code{x}
to compute non-parametric statistics on. Useful when a subset
of quality-assured data is available. Default action is to
ignore NA values.}

\item{threshold}{A length-two vector identifying
thresholds for "mild" and "extreme" outliers.}

\item{return.score}{if \code{TRUE}, return the numeric outlier score.
If FALSE, return a label classifying the observations as one of
"not outlier", "mild outlier", or "extreme outlier".}
}
\description{
Performs a Chi-squared test for outliers.
}
\details{
the values of \code{threshold} identify the quantiles of the
Chi-squared distribution used to identify mild and extreme outliers.
Default values are 0.9 for "mild" outliers and 0.95 for "extreme"
outliers.
}
