% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.r
\name{daily_tidal_mean}
\alias{daily_tidal_mean}
\title{Daily Tidal Mean}
\usage{
daily_tidal_mean(
  x,
  y,
  tide = c("high", "low"),
  smooth = TRUE,
  neighborhood = 24,
  family = "symmetric",
  degree = 2,
  ...,
  plot = FALSE
)
}
\arguments{
\item{x}{A vector of values from a single day.}

\item{y}{A vector of water surface elevations from a single day.}

\item{tide}{Return the high-tide ("high") or low-tide ("low")
average.}

\item{smooth}{If \code{TRUE}, A loess smoother will be applied to the
data prior to detecting the water surface elevation minimums
or maximums.}

\item{neighborhood}{The neighborhood (number of points) used to
determine the \code{span} argument to \code{\link[stats:loess]{stats::loess()}}. A time
window of 6 hours was found to work well with tidal data, which
corresponds to 24 points for 15-minute data or 6 points for
hourly data.}

\item{family}{The \code{family} argument to \code{\link[stats:loess]{stats::loess()}}. The
default family "symmetric" was found to work well with 15-minute
tidal data.}

\item{degree}{The \code{degree} argument to \code{\link[stats:loess]{stats::loess()}}. The
default degree of 2 was found to work well with 15-minute
tidal data.}

\item{...}{Additional arguments to \code{\link[stats:loess]{stats::loess()}}.Note that
the value of \code{span} is computed from the argument \code{neighborhood}.}

\item{plot}{If \code{TRUE}, printdiagnostic plots of the loess smoothing.}
}
\value{
The daily tidal mean of \code{x}, i.e. the average of high-
or low-tide values for the day.
}
\description{
Compute the daily low-tide or high-tide average.
}
