% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapfill.r
\name{gapfill_kalman}
\alias{gapfill_kalman}
\title{ARIMIA and Kalman Filter Gap Fill}
\usage{
gapfill_kalman(x, mask = is.na(x), ...)
}
\arguments{
\item{x}{A vector of data with missing values.}

\item{mask}{A logical mask that identifies a subgroup of \code{x}
to compute fill values for. Useful when a subset
of quality-assured data is available. Default action is to
only fill NA values.}

\item{...}{Additional arguments to \code{\link[forecast:auto.arima]{forecast::auto.arima()}}.}
}
\value{
The original vector x with missing values imputed.
}
\description{
Fill gaps in a timeseries by fitting an ARIMA model and applying
a Kalman filter.
}
