% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{iforest_outliers}
\alias{iforest_outliers}
\title{Isolation Forest outlier detection}
\usage{
iforest_outliers(
  x,
  mask = !is.na(x),
  sample.size = length(mask),
  threshold = c(0.8, 0.9),
  return.score = FALSE,
  ...
)
}
\arguments{
\item{x}{A vector of data.}

\item{mask}{A logical mask that identifies a subgroup of \code{x}
to compute non-parametric statistics on. Useful when a subset
of quality-assured data is available. Default action is to
remove NA values.}

\item{threshold}{A length-two vector identifying
thresholds for "mild" and "extreme" outliers.}

\item{return.score}{if \code{TRUE}, return the numeric outlier score.
If FALSE, return a label classifying the observations as one of
"not outlier", "mild outlier", or "extreme outlier".}

\item{...}{Additional arguments to \code{ranger::ranger} (used by
\code{solitude::isolationForest}).}
}
\description{
Performs outlier detection using an Isolation Forest.
}
\details{
the values of \code{threshold} identify mild and extreme\
outliers based on the Isolation Forest score in the range \verb{[0,1]}.
Default values are 0.8 for "mild" outliers and 0.9 for "extreme"
outliers.
}
