% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers-multivar.r
\name{outlier_lof}
\alias{outlier_lof}
\title{LOF outlier detection}
\usage{
outlier_lof(
  xs,
  mask = !Reduce("|", lapply(xs, is.na)),
  threshold = c(1.5, 2),
  return.score = FALSE,
  ...
)
}
\arguments{
\item{xs}{A dataframe or list of vectors
(which will be coerced to a numeric matrix).}

\item{mask}{A logical vector that defines which values in \code{x}
will used when computing statistics. Useful when a subset
of quality-assured data is available. Default mask is non-\code{NA}
Values.}

\item{threshold}{A length-two vector identifying
thresholds for "mild" and "extreme" outliers.}

\item{return.score}{if \code{TRUE}, return the numeric outlier score.
If \code{FALSE}, return an ordered factor classifying the observations
as one of "not outlier" (1), "mild outlier" (2), or
"extreme outlier" (3).}

\item{...}{Additional arguments to \code{dbscan::lof}, namely
\code{k}.}
}
\description{
Performs outlier detection using Local Outlier Factor algorithm.
}
\details{
the values of \code{threshold} identify mild and extreme
outliers based on the LOF score. Scores significantly larger
than 1 indicate outliers. Default values are 1.5 for "mild"
outliers and 2.0 for "extreme" outliers.
}
\examples{
x = seq(0, 34, by = 0.25)*pi
noise = rlnorm(length(x), meanlog = 1, sdlog = 3)
y=sin(x) + noise
mask = noise < 1

if (require(dbscan)) {
  outlier_lof(list(y))
  outlier_lof(list(x, y), mask)
  outlier_lof(list(x, y), mask, threshold = c(1, 2))
  outlier_lof(list(x, y), return.score = TRUE)
}

}
