% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.r
\name{outlier_tscore}
\alias{outlier_tscore}
\title{t-score test for outliers}
\usage{
outlier_tscore(
  x,
  mask = !is.na(x),
  threshold = c(0.9, 0.95),
  return.score = FALSE
)
}
\arguments{
\item{x}{A vector of data.}

\item{mask}{A logical vector that defines which values in \code{x}
will used when computing statistics. Useful when a subset
of quality-assured data is available. Default mask is non-\code{NA}
Values.}

\item{threshold}{A length-two vector identifying
thresholds for "mild" and "extreme" outliers.}

\item{return.score}{if \code{TRUE}, return the numeric outlier score.
If \code{FALSE}, return an ordered factor classifying the observations
as one of "not outlier" (1), "mild outlier" (2), or
"extreme outlier" (3).}
}
\description{
Performs a t-score test for outliers.
}
\details{
The values of \code{threshold} identify the quantiles of the
t-distribution used to identify mild and extreme outliers.
Default values are 0.9 for "mild" outliers and 0.95 for "extreme"
outliers.

The t-score is equivalent to the z-score for sample
sizes greater than 30.
}
\examples{
x = seq(0, 34, by = 0.25)*pi
noise = rnorm(length(x), mean = 1, sd = 3)
y = sin(x) + noise
mask = noise < 1

outlier_tscore(y)
outlier_tscore(y, mask)
outlier_tscore(y, mask, threshold = c(0.8, 0.9))
outlier_tscore(y, return.score = TRUE)

}
