% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtqc.r
\name{range_test}
\alias{range_test}
\title{Range Test}
\usage{
range_test(x, sensor.range, user.range)
}
\arguments{
\item{x}{A vector of values.}

\item{sensor.range}{A length-2 numeric vector identifying the sensor
measurement range.}

\item{user.range}{A length-2 numeric vector identifying a reasonable
measurement range for the provided data \code{x}. Typically specific to
location and/or climate and based on expert judgment.}
}
\value{
An integer vector of test flags.
}
\description{
Perform a range test. See
https://cdn.ioos.noaa.gov/media/2017/12/qartod_temperature_salinity_manual.pdf
for more information.
}
\examples{
fake.data = c(rnorm(3, 10), rep(c(-5, 30), each =  3), rep(c(5, 20), each = 3))
range_test(fake.data, c(0, 25), c(7, 15))

}
