% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtqc.r
\name{rtqc_attenuation}
\alias{rtqc_attenuation}
\title{Attenuation Test}
\usage{
rtqc_attenuation(x, threshold, n.obs, ...)
}
\arguments{
\item{x}{A vector of values.}

\item{threshold}{A length-2 vector specifying "suspect" and "fail"
thresholds for the minimum standard deviation of a set of
observations.}

\item{n.obs}{The number of observations, including the current
observation, to use for calculating standard deviation.}

\item{...}{Other arguments to pass to \code{\link[stats:sd]{stats::sd()}}, i.e.,
the argument \code{na.rm}.}
}
\value{
An ordered factor of test flags of same length as \code{x}.
}
\description{
Perform a signal attenuation test. For more information see
U.S. Integrated Ocean Observing System, 2015. Manual for Real-Time
Quality Control of In-situ Temperature and Salinity Data Version
2.0: A Guide to Quality Control and Quality Assurance of In-situ
Temperature and Salinity Observations. 56 pp.
DOI: \href{https://doi.org/10.7289/V5V40SD4}{10.7289/V5V40SD4}.
}
\examples{
fake.data = sin(seq(0, 10, by = 0.1) * pi) * seq(10, 0, by = -0.1)
rtqc_attenuation(fake.data, c(1, 0.5), 10)

}
