% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtqc.r
\name{rtqc_range}
\alias{rtqc_range}
\title{Range Test}
\usage{
rtqc_range(x, user.range, sensor.range)
}
\arguments{
\item{x}{A vector of values.}

\item{user.range}{A length-2 numeric vector identifying a reasonable
measurement range for the provided data \code{x}. Typically specific to
location and/or climate and based on expert judgment.}

\item{sensor.range}{A length-2 numeric vector identifying the sensor
measurement range.}
}
\value{
An ordered factor of test flags of same length as \code{x}.
}
\description{
Perform a range test. For more information see
U.S. Integrated Ocean Observing System, 2015. Manual for Real-Time
Quality Control of In-situ Temperature and Salinity Data Version
2.0: A Guide to Quality Control and Quality Assurance of In-situ
Temperature and Salinity Observations. 56 pp.
DOI: \href{https://doi.org/10.7289/V5V40SD4}{10.7289/V5V40SD4}.
}
\examples{
fake.data = c(rnorm(3, 10), rep(c(-5, 30), each =  3),
  rep(c(5, 20), each = 3))
rtqc_range(fake.data, c(7, 15), c(0, 25))

}
