% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.r
\name{smooth_godin}
\alias{smooth_godin}
\title{Godin Smoother}
\usage{
smooth_godin(x, increment, kind = c("mean", "max", "min", "median"), ...)
}
\arguments{
\item{x}{A vector of values.}

\item{increment}{A character string defining the expected time
increment. this consists of two parts: a numeric value defining
the size of the increment, and the units of the increment.
Acceptable units are "secs", "mins", "hours", "days", or "weeks".}

\item{kind}{what values to filter. Default is the original Godin
filter which averages the data. Other options are to use the \code{max}
or \code{min} values of each window.}

\item{...}{Additional arguments to \code{mean()}, \code{max()}, \code{min()},
or \code{median()}, depending on the argument \code{kind}.}
}
\value{
A vector of same length as x.
}
\description{
Apply a Godin filter to a vector of data.
}
\details{
The Godin filter is defined
\deqn{F = \frac{A_{24}^2}{24^2}\frac{A_{25}}{25}}
}
